fixed firebase error and organized files
